requirements.txt
# MiniMe Core Dependencies
# Install with: pip install -r requirements.txt

typer[all]>=0.9.0
pydantic>=2.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-frontmatter>=0.5.0
numpy>=1.24.0

# Visualization dependencies
plotly>=5.0.0
networkx>=3.0
matplotlib>=3.5.0

# Vector search dependencies
faiss-cpu>=1.7.4

